requirements.txt
pandas~=1.1.5
numpy~=1.19.5
py4j~=0.10.9.3
pytz~=2022.1
cloudpickle~=1.2.2
setuptools~=62.1.0
Cython~=0.29.24
torch~=1.11.0
onnx~=1.11.0
Pillow~=9.1.0
scikit-learn~=1.0.2
torchvision~=0.12.0
requests~=2.27.1
onnxruntime~=1.11.0